api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
openpyxl==3.1.2
python-dotenv==1.0.0
googlemaps==4.10.0
requests==2.31.0
pydantic==2.5.2
python-multipart==0.0.6
supabase==2.3.5
sqlalchemy==2.0.25
asyncpg==0.29.0
mangum==0.17.0
PyJWT==2.8.0 